Struct board_ecs::entities::Entities

pub struct Entities {
    alive: BitSetVec,
    generation: Vec<u32>,
    killed: Vec<Entity>,
    next_id: usize,
    has_deleted: bool,
}

Holds a list of alive entities.

It also holds a list of entities that were recently killed, which allows to remove components of deleted entities at the end of a game frame.

Fields
alive: BitSetVec

Bitset containing all living entities

generation: Vec<u32>killed: Vec<Entity>next_id: usizehas_deleted: bool

helps to know if we should directly append after next_id or if we should look through the bitset.

Implementations

impl Entities

pub fn create(&mut self) -> Entity

Creates a new Entity and returns it.

This function will not reuse the index of an entity that is still in the killed entities.

pub fn is_alive(&self, entity: Entity) -> bool

Checks if the Entity is still alive.

Returns true if it is alive. Returns false if it has been killed.

pub fn kill(&mut self, entity: Entity)

Kill an entity.

pub fn killed(&self) -> &Vec<Entity>

Returns entities in the killed list.

pub fn clear_killed(&mut self)

Clears the killed entity list.

pub fn bitset(&self) -> &BitSetVec

Returns a bitset where each index where the bit is set to 1 indicates the index of an alive entity.

Useful for joining over Entity and ComponentStore<T> at the same time.

pub fn iter_with_bitset(&self, bitset: Rc<BitSetVec>) -> EntityIterator<'_>

Iterates over entities using the provided bitset.