Module board_ecs::prelude

The prelude.

Re-exports

pub use crate::join;

pub use crate::EcsData;

pub use crate::RawFns;

pub use crate::TypedEcsData;

pub use crate::bitset::*;

pub use crate::components::*;

pub use crate::dispatcher::*;

pub use crate::entities::*;

pub use crate::resources::*;

pub use crate::system::*;

pub use crate::uuid::*;

Macros

join

The join macro makes it very easy to iterate over multiple components of the same Entity at once.

Structs

AtomicRef

A wrapper type for an immutably borrowed value from an AtomicRefCell<T>.

AtomicRefCell

A threadsafe analogue to RefCell.

AtomicRefMut

A wrapper type for a mutably borrowed value from an AtomicRefCell<T>.

BorrowError

An error returned by AtomicRefCell::try_borrow.

BorrowMutError

An error returned by AtomicRefCell::try_borrow_mut.

Uuid

A Universally Unique Identifier (UUID).

World

The World is simply a collection of Resources, and ComponentStores.

Enums

EcsError

The types of errors used throughout the ECS.

Traits

BitSet

The BitSet API.

TypeUuid

Provides a statically defined UUID for a Rust type.

Type Definitions

SystemResult

The result of a System’s execution.

Derive Macros

Deref

Implements Deref for single-item structs. This is especially useful when utilizing the newtype pattern.

DerefMut

Implements DerefMut for single-item structs. This is especially useful when utilizing the newtype pattern.

TypeUuid