Struct board_ecs::resources::Resources

pub struct Resources {
    untyped: UntypedResources,
    type_ids: UuidMap<TypeId>,
}

A collection of resources.

Resources is essentially a type-map

Fields
untyped: UntypedResourcestype_ids: UuidMap<TypeId>
Implementations

impl Resources

pub fn new() -> Self

Create an empty Resources.

pub fn init<T: TypedEcsData + Default>(&mut self)

Initialize a resource of type T by inserting it’s default value.

pub fn insert<T: TypedEcsData>(&mut self, resource: T)

Insert a resource.

Panics

Panics if you try to insert a Rust type with a different TypeId, but the same TypeUuid as another resource in the store.

pub fn try_insert<T: TypedEcsData>(
&mut self,
resource: T
) -> Result<(), EcsError>

Try to insert a resource.

Errors

Errors if you try to insert a Rust type with a different TypeId, but the same TypeUuid as another resource in the store.

pub fn get<T: TypedEcsData>(&self) -> AtomicResource<T>

Get a resource handle from the store.

This is not the resource itself, but a handle, may be cloned cheaply.

In order to access the resource you must call borrow() or borrow_mut() on the returned value.

Panics

Panics if the resource does not exist in the store.

pub fn contains<T: TypedEcsData>(&self) -> bool

Check whether or not a resource is in the store.

See get()

pub fn try_get<T: TypedEcsData>(&self) -> Option<AtomicResource<T>>

Gets a resource handle from the store if it exists.

pub fn untyped(&self) -> &UntypedResources

Borrow the underlying UntypedResources store.

pub fn untyped_mut(&mut self) -> &mut UntypedResources

Mutably borrow the underlying UntypedResources store.

pub fn into_untyped(self) -> UntypedResources

Consume Resources and extract the underlying UntypedResources.